requirements.txt
exceptiongroup==1.2.2
greenlet==3.1.1
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.9
pyodbc==5.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
